Refactoring: dedicated file for URL functions
