+++ source/connectors/github/helpers.go
@@ -17,51 +17,10 @@
 package github
 
 import (
-	"net/url"
-	"path"
-
 	"github.com/artem-sidorenko/chagen/source/connectors/helpers"
 )
 
 // formatErrorCode formats the error message for this connector
 func formatErrorCode(query string, err error) error {
 	return helpers.FormatErrorCode("GitHub", query, err)
-}
-
-// getTagURL returns the URL for a given tag.
-// If alwaysUseReleaseURL is true: URL is provided for release page,
-// even if it does not exist yet
-func (c *Connector) getTagURL(tagName string, alwaysUseReleaseURL bool) (string, error) {
-	release, resp, err := c.client.Repositories.GetReleaseByTag(c.context, c.Owner, c.Repo, tagName)
-	if err != nil {
-		// no release was found for this tag, this is no error for us
-		if resp.StatusCode != 404 {
-			return "", formatErrorCode("getTagURL", err)
-		}
-	}
-
-	// if GitHub release for this tag was found -> use it
-	// generate otherwise a link to the git tag view in the file tree
-	var tagURL string
-	if release != nil { // we got real release URL, use it
-		tagURL = release.GetHTMLURL()
-	} else { // build own URL
-		u, err := url.Parse(c.ProjectURL)
-		if err != nil {
-			return "", err
-		}
-
-		if alwaysUseReleaseURL { // try to build own release url
-			u.Path = path.Join(u.Path, "/releases/"+tagName)
-		} else { // build tag url
-			u.Path = path.Join(u.Path, "/tree/"+tagName)
-		}
-		tagURL = u.String()
-	}
-	return tagURL, nil
-}
-
-// GetNewTagURL returns the URL for a new tag, which does not exist yet
-func (c *Connector) GetNewTagURL(TagName string) (string, error) {
-	return c.getTagURL(TagName, c.NewTagUseReleaseURL)
 }
